organized main.cpp

+++ include/Line.h
@@ -6,11 +6,11 @@
 
 class Line {
 public:
-    float x1, y1, x2, y2;
+    int x1, y1, x2, y2;
     color c{};
 
-    Line(const float x1, const float y1, const float x2, const float y2, const color c);
-    Line(const float x1, const float y1, const float x2, const float y2);
+    Line(const int x1, const int y1, const int x2, const int y2, const color c);
+    Line(const int x1, const int y1, const int x2, const int y2);
     void display() const;
 };
